% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkyPloid.R
\name{funkyPloid}
\alias{funkyPloid}
\title{calculate LLR's for a group of samples and a given set of ploidy values}
\usage{
funkyPloid(counts, counts_alt = NULL, ploidy, h = NULL, eps = NULL,
  maxIter = 10000, maxDiff = 0.001)
}
\description{
calculate LLR's for a group of samples and a given set of ploidy values
}
