% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoProps.R
\name{genoProps}
\alias{genoProps}
\title{calculate proportions of each genotype category for a group of samples and a given ploidy value}
\usage{
genoProps(counts, counts_alt = NULL, ploidy, h = NULL, eps = NULL,
  maxIter = 10000, maxDiff = 0.001)
}
\description{
calculate proportions of each genotype category for a group of samples and a given ploidy value
}
