% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPloidySamples.R
\name{rPloidySamples}
\alias{rPloidySamples}
\title{simulate read counts for random samples}
\usage{
rPloidySamples(
  nSamps,
  reads,
  truePloidy,
  alpha,
  eps = NULL,
  h = NULL,
  genotypeCounts,
  genotypePropsAreKnown = FALSE
)
}
\arguments{
\item{nSamps}{number of samples to simulate}

\item{reads}{number of reads per sample}

\item{truePloidy}{the true ploidy, as an integer, to simulate}

\item{alpha}{the alpha values of a Dirichlet distribution describing the
proportion of reads for each locus}

\item{eps}{Either \code{NULL}, or a numeric vector of values for the error rate per read for each locus in the same
order that the loci are ordered in \code{alpha}}

\item{h}{Either \code{NULL}, or a numeric vector of h values for each locus in the same order that the
loci are ordered in \code{alpha}}

\item{genotypeCounts}{a matrix with each row representing a locus (in the same order as \code{alpha}),
and each column representing a genotype with column 1 being 0 copies of the reference allele, column 2
being 1 copy of the reference, ..., column \code{truePloidy + 1} being \code{truePloidy} copies of the reference}

\item{genotypePropsAreKnown}{boolean indicating whether to threat \code{genotypeCounts} as true proportions,
or the observed number of genotypes in a sample}
}
\description{
for nSamps samples, read counts are simulated by:
  1. If genotypePropsAreKnown is \code{FALSE}, draw genotype proportions from a Dirichlet posterior calculated
    using genotypeCounts (treated as a multinomial) and a uniform Dirichlet prior. If
    genotypePropsAreKnown is \code{TRUE}, treat genotypeCounts as the true genotype proportions.
  2. Draw genotypes from categorical distributions using the genotype proportions from step 1.
  3. Draw read counts per locus from a Dirichlet-multinomial with \code{alpha} describing the Dirichlet
    and \code{reads} total reads
  4. Draw reads of reference and alternative alleles as a binomial using the probability calculated from
    the allele dosage of the genotypes drawn in step 2, \code{eps}, and \code{h}
}
